Extract maven dependencies out of app-engine into optional maven module
