+++ .idea/compiler.xml
@@ -63,6 +63,8 @@
       <module name="java-ultimate_test" target="1.8" />
       <module name="java_main" target="1.8" />
       <module name="java_test" target="1.8" />
+      <module name="maven_main" target="1.8" />
+      <module name="maven_test" target="1.8" />
       <module name="stackdriver-debugger_main" target="1.8" />
       <module name="stackdriver-debugger_test" target="1.8" />
       <module name="ultimate_main" target="1.8" />
