+++ app-engine/java/gradle/build.gradle
@@ -22,7 +22,7 @@ sourceSets.test.java.srcDirs = ['testSrc']
 sourceSets.test.resources.srcDirs = ['testResources']
 
 intellij {
-    plugins 'Groovy', 'gradle', 'git4idea', 'properties', 'junit', 'yaml', 'maven'
+    plugins 'gradle'
 }
 
 dependencies {
