+++ app-engine/java/maven/build.gradle
@@ -22,18 +22,13 @@ sourceSets.test.java.srcDirs = ['testSrc']
 sourceSets.test.resources.srcDirs = ['testResources']
 
 intellij {
-    plugins 'Groovy', 'gradle', 'git4idea', 'properties','maven', 'yaml'
+    plugins 'maven'
 }
 
 dependencies {
+    compile project(':app-engine:java')
     compile(project(':google-cloud-core'))
-    compile(project(':app-engine:java'))
-    compile(project(':app-engine:java:maven'))
-    compile 'com.google.apis:google-api-services-servicemanagement:v1-rev374-1.22.0'
-    compile ('com.google.cloud.tools:appengine-plugins-core:' + toolsLibVersion) {
-        exclude group: 'com.google.guava', module: 'guava'
-        exclude group: 'org.yaml', module: 'snakeyaml'
-    }
+
     compileOnly 'com.google.auto.value:auto-value:1.4.1'
     apt         'com.google.auto.value:auto-value:1.4.1'
