+++ app-engine/java/maven/resources/META-INF/app-engine-java-maven.xml
@@ -1,0 +1,33 @@
+<!--
+  ~ Copyright 2018 Google Inc. All Rights Reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<idea-plugin>
+
+    <extensions defaultExtensionNs="com.intellij">
+        <applicationService serviceImplementation="com.google.cloud.tools.intellij.appengine.java.maven.project.MavenProjectService"/>
+
+        <remoteServer.deploymentSource.type
+                implementation="com.google.cloud.tools.intellij.appengine.java.maven.MavenBuildDeploymentSourceType"/>
+    </extensions>
+
+    <extensions defaultExtensionNs="com.google.gct.core">
+        <appEngineDeploymentSourceProvider
+                implementation="com.google.cloud.tools.intellij.appengine.java.maven.AppEngineMavenDeploymentSourceProvider"/>
+
+        <appEngineStandardLibraryManager
+                implementation="com.google.cloud.tools.intellij.appengine.java.maven.AppEngineStandardMavenLibraryManager"/>
+    </extensions>
+</idea-plugin>
