+++ app-engine/java/maven/src/com.google.cloud.tools.intellij.appengine.java.maven/AppEngineMavenDeploymentSourceProvider.java
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.java.cloud;
+package com.google.cloud.tools.intellij.appengine.java.maven;
 
+import com.google.cloud.tools.intellij.appengine.java.cloud.AppEngineDeploymentSourceProvider;
+import com.google.cloud.tools.intellij.appengine.java.cloud.AppEngineEnvironment;
 import com.google.cloud.tools.intellij.appengine.java.facet.flexible.AppEngineFlexibleFacet;
 import com.google.cloud.tools.intellij.appengine.java.facet.standard.AppEngineStandardFacet;
 import com.google.cloud.tools.intellij.appengine.java.project.AppEngineProjectService;
-import com.google.cloud.tools.intellij.appengine.java.project.MavenProjectService;
+import com.google.cloud.tools.intellij.appengine.java.maven.project.MavenProjectService;
 import com.google.common.collect.Lists;
 import com.intellij.openapi.module.JavaModuleType;
 import com.intellij.openapi.module.Module;
