+++ app-engine/java/src/com/google/cloud/tools/intellij/appengine/java/cloud/BuildDeploymentSourceType.java
@@ -43,10 +43,10 @@ import org.jetbrains.annotations.Nullable;
 public abstract class BuildDeploymentSourceType
     extends DeploymentSourceType<ModuleDeploymentSource> {
 
-  static final String CLOUD_PROJECT_ATTRIBUTE = "project";
-  static final String APP_ENGINE_VERSION_ATTRIBUTE = "version";
+  public static final String CLOUD_PROJECT_ATTRIBUTE = "project";
+  public static final String APP_ENGINE_VERSION_ATTRIBUTE = "version";
 
-  BuildDeploymentSourceType(@NotNull String id) {
+  public BuildDeploymentSourceType(@NotNull String id) {
     super(id);
   }
