+++ google-cloud-tools-plugin/resources/META-INF/plugin.xml
@@ -50,6 +50,7 @@
   <xi:include href="/META-INF/google-cloud-core.xml" xpointer="xpointer(/idea-plugin/*)"/>
   <xi:include href="/META-INF/google-cloud-storage.xml" xpointer="xpointer(/idea-plugin/*)"/>
   <xi:include href="/META-INF/google-cloud-repos.xml" xpointer="xpointer(/idea-plugin/*)"/>
+  <xi:include href="/META-INF/google-cloud-maven.xml" xpointer="xpointer(/idea-plugin/*)"/>
 
   <depends>com.intellij.modules.lang</depends>
   <depends>Git4Idea</depends>
