+++ settings.gradle
@@ -21,6 +21,7 @@ include 'google-cloud-repos'
 include 'google-cloud-apis'
 include 'stackdriver-debugger'
 include 'app-engine:java'
+include 'app-engine:java:maven'
 include 'app-engine:java:gradle'
 include 'app-engine:java:ultimate'
 include 'google-cloud-sdk'
